Add pagination logic to a pagination service
