+++ src/app/components/app/app.component.css
@@ -17,7 +17,7 @@
     left: 0;
     width: 250px;
     height: 55px;
-    position: relative;
+    position: fixed;
     box-shadow: -5px 2px 6px 2px rgb(71, 74, 103);
     z-index: 1;
 }
